+++ HISTORY.rst
@@ -27,6 +27,9 @@ latest (unreleased)
 
 **Bugfixes**
 
+* Addresses imported in orders are copied to new addresses. Now they are
+  imported as inactive.
+
 **Build**
 
 **Documentation**